+++ src/components/BookDetail.jsx
@@ -106,7 +106,7 @@ export default function BookDetail({ books, user, addToast }) {
               <img src={book.cover} alt={book.title}
                 style={{ width: "100%", height: "100%", objectFit: "cover" }} />
             ) : (
-              <BookPlaceholder title={book.title} />
+              <BookPlaceholder title={book.title} author={book.author} genre={book.genre} />
             )}
           </div>
           {!book.cover && (