+++ src/components/Recommendations.jsx
@@ -77,7 +77,7 @@ export default function Recommendations({ books, addToast }) {
                   <img src={rec.cover} alt={rec.title}
                     style={{ width: "100%", height: "100%", objectFit: "cover" }} />
                 ) : (
-                  <BookPlaceholder title={rec.title} />
+                  <BookPlaceholder title={rec.title} author={rec.author} genre={rec.genre} />
                 )}
               </div>
               <div style={{ flex: 1 }}>